Added functionality to download models and chat
If the app can access the internet, the model Phi3 from the Microsoft HF repository is downloaded if
no model is already present in the application.

Started to work on the chat ability of the model. It execute one request on a separated thread,
allowing for the response from the model in background.

Signed-off-by: Dennis Dosso <dennis.dosso@secomind.com>